Squashed commit of the following:
commit eb8614b088e2957f2bc0f15bda65b6f85fc5d54e
Author: Rafael Singer <rafaelsinger66@gmail.com>
Date:   Thu Apr 24 01:03:17 2025 -0400

    bug fixes with DirectoryContent

commit e509ef7ab43b8828676a32ecf3084cfd1f71f7d4
Merge: 689df2f df18a5f
Author: Hunter S <67876717+hunterschep@users.noreply.github.com>
Date:   Thu Apr 24 00:50:55 2025 -0400

    Merge branch 'staging' of https://github.com/rafaelsinger/flock into staging

commit 689df2fdfaff32d32de98d5445090c0f66ec9635
Author: Hunter S <67876717+hunterschep@users.noreply.github.com>
Date:   Thu Apr 24 00:50:46 2025 -0400

    class year dropdown

commit df18a5fe82565a692d0991d33220bbb21c0e3284
Author: Rafael Singer <rafaelsinger66@gmail.com>
Date:   Thu Apr 24 00:49:04 2025 -0400

    increase limit for top destinations

commit a7b85936e1e05e9a9ccb67d2a6c1c4cafe8bdae7
Author: Rafael Singer <rafaelsinger66@gmail.com>
Date:   Thu Apr 24 00:43:33 2025 -0400

    fix bug on onboarding flow

commit 7f09779d94ead7a74072127721651f2dcf7f4ac0
Author: Rafael Singer <rafaelsinger66@gmail.com>
Date:   Thu Apr 24 00:27:43 2025 -0400

    ui changes to onboarding and refactoring to use constant for current class year

commit 4e385a93339b517e4155dfcdeffec9a645938ab0
Author: Hunter S <67876717+hunterschep@users.noreply.github.com>
Date:   Thu Apr 24 00:27:09 2025 -0400

    make filter more modular

commit 1f28b03a46bb5d5c71596487d5834665da74984a
Author: Hunter S <67876717+hunterschep@users.noreply.github.com>
Date:   Thu Apr 24 00:21:59 2025 -0400

    inital button fix

commit 6814a778f4415a14d9db12be65af5745da7bd068
Author: Rafael Singer <rafaelsinger66@gmail.com>
Date:   Wed Apr 23 23:59:52 2025 -0400

    add fix for washington dc edge case

commit 730d61c8cbc41935ae26d732c14b32862660a5ed
Author: Hunter S <67876717+hunterschep@users.noreply.github.com>
Date:   Wed Apr 23 23:34:04 2025 -0400

    fix industry passback to be a string not an object

    allows save button to work

commit 1859e07499fa2663559db8fabee21c1aab6f748f
Author: Rafael Singer <rafaelsinger66@gmail.com>
Date:   Wed Apr 23 23:19:42 2025 -0400

    try to fix excessive loading on conversation page

commit 56c97bbfb36c262bf2ec2771ae314c193820ce5c
Author: Rafael Singer <rafaelsinger66@gmail.com>
Date:   Wed Apr 23 21:28:09 2025 -0400

    fix small issue on edit page

commit 4c817cb8613ebccd5561d0ce2e7de0ec4989f354
Author: Rafael Singer <rafaelsinger66@gmail.com>
Date:   Wed Apr 23 21:12:51 2025 -0400

    add class year to top destinations

commit 24d4717cdc8a42e04683de9da6363ee882ead79d
Author: Rafael Singer <rafaelsinger66@gmail.com>
Date:   Wed Apr 23 20:53:43 2025 -0400

    add class year to locations endpoint

commit 540dbbcad3e675996669b0f019076e93fdd59f56
Author: Rafael Singer <rafaelsinger66@gmail.com>
Date:   Wed Apr 23 17:38:12 2025 -0400

    add zod schemas

commit d6af7b6507bf500e44213759f3da34a82850a2e3
Author: Rafael Singer <rafaelsinger66@gmail.com>
Date:   Mon Apr 21 21:34:18 2025 -0400

    fix minor build issue

commit a89326b8174760a7ead1ed46c1be3d0bac70c250
Merge: d793ba0 498859b
Author: Hunter M Scheppat <67876717+hunterschep@users.noreply.github.com>
Date:   Mon Apr 21 01:33:05 2025 -0700

    Merge pull request #5 from rafaelsinger/ui-updates

    First UI updates

commit 498859b112e988ff02d74b3c82beda9a9fc06141
Author: Hunter S <67876717+hunterschep@users.noreply.github.com>
Date:   Mon Apr 21 01:27:51 2025 -0700

    improve onboarding on mobile

commit e72ffd8b5e36ad6fb4891eaae488b1b35e6a385b
Author: Hunter S <67876717+hunterschep@users.noreply.github.com>
Date:   Mon Apr 21 00:50:07 2025 -0700

    make user cards cleaner

commit 3b9e067b6409442320b209753d75a0206d9fff16
Author: Hunter S <67876717+hunterschep@users.noreply.github.com>
Date:   Mon Apr 21 00:19:29 2025 -0700

    improve messaging ui

commit b96a533aad6ff9065c55b653a2eec4da2b187ad9
Author: Hunter S <67876717+hunterschep@users.noreply.github.com>
Date:   Mon Apr 21 00:16:48 2025 -0700

    clean up messaging ui

commit 58242e78e9071a9a6103bfb7a418ddbc95e044f7
Author: Hunter S <67876717+hunterschep@users.noreply.github.com>
Date:   Mon Apr 21 00:12:07 2025 -0700

    fix card overflow

commit 4578bf3d991569a2d6f8a13f3a759bdc9afc240b
Author: Hunter S <67876717+hunterschep@users.noreply.github.com>
Date:   Mon Apr 21 00:07:10 2025 -0700

    add cursor pointer on hover

commit 2a44317787cdee0d9f0f21ee3bf56fdb7d1e7e84
Author: Hunter S <67876717+hunterschep@users.noreply.github.com>
Date:   Sun Apr 20 23:26:58 2025 -0700

    dont allow empty form submission

commit 828bf726fe280a68c1cbad3ed098534c92b1d177
Author: Hunter S <67876717+hunterschep@users.noreply.github.com>
Date:   Sun Apr 20 23:22:50 2025 -0700

    constrain user input to types on profile edit and tweak delete button

commit f1cd90b49b5c59c471f27011b06f4bbc626fee53
Author: Hunter S <67876717+hunterschep@users.noreply.github.com>
Date:   Sun Apr 20 23:10:09 2025 -0700

    remove option to select a bachelors

    users are postgrad only for this path

commit fa920d0d4a0401dc77ac6739425574253c9a5888
Author: Hunter S <67876717+hunterschep@users.noreply.github.com>
Date:   Sun Apr 20 21:53:51 2025 -0700

    fix more copy to support interns and grads

commit cd514a436dbd12652fdc25888e77970f87ba3717
Author: Hunter S <67876717+hunterschep@users.noreply.github.com>
Date:   Sun Apr 20 21:41:20 2025 -0700

    fix copy for interns + grads

commit d793ba09a41c6e07091f9abdc08a9b30fa05766b
Author: Rafael Singer <rafaelsinger66@gmail.com>
Date:   Sun Apr 20 23:19:00 2025 -0400

    fix build errors

commit 16763567f09d065c261dfa788dee96af6ff0353b
Author: Rafael Singer <rafaelsinger66@gmail.com>
Date:   Sun Apr 20 23:14:31 2025 -0400

    revert internship postgradtype migration since prisma gets mad

commit d15d31fdeb87c7140fd7ca6e45488dd3f403b24c
Author: Rafael Singer <rafaelsinger66@gmail.com>
Date:   Sun Apr 20 22:58:29 2025 -0400

    merge into staging

commit 4c35e81cd0071614324ccdeb85d15c9af1c8e689
Author: Rafael Singer <rafaelsinger66@gmail.com>
Date:   Sun Apr 20 18:59:21 2025 -0400

    make migration idempotent

commit ef1a4069129a10f0dba2ae4e6e94768839add6e0
Author: Rafael Singer <rafaelsinger66@gmail.com>
Date:   Sun Apr 20 18:54:52 2025 -0400

    add intern support

commit e960b18a1376cad9b869917b89f43ea94555b953
Author: Rafael Singer <rafaelsinger66@gmail.com>
Date:   Sat Apr 19 13:43:32 2025 -0500

    small filter panel fix

commit 0549f534be30461df6eea382a608d9e3c837bcb7
Merge: acb775d 39f5741
Author: Hunter M Scheppat <67876717+hunterschep@users.noreply.github.com>
Date:   Fri Apr 18 21:17:59 2025 -0700

    Merge pull request #4 from rafaelsinger/hotfix-map-discrepancy

    Fix map discrepancy

commit 39f574154ab2f389af81954f35582ccc52d6348f
Author: Hunter S <67876717+hunterschep@users.noreply.github.com>
Date:   Fri Apr 18 21:15:37 2025 -0700

    Update route.ts

commit acb775db883a634066610b59431509714783420f
Merge: bc48aa3 67a2d2e
Author: Rafael Singer <rafaelsinger66@gmail.com>
Date:   Fri Apr 18 19:54:30 2025 -0500

    Merge branch 'main' into staging

commit bc48aa3f0f3588a9aebb2d548967e40595815e69
Merge: 78b3bbe ea0ec8e
Author: Rafael Singer <rafaelsinger66@gmail.com>
Date:   Thu Apr 17 23:22:40 2025 -0500

    Merge branch 'main' into staging

commit 78b3bbe8f79c3f9a881b833baf5d84e0986a059e
Author: Rafael Singer <rafaelsinger66@gmail.com>
Date:   Thu Apr 17 14:48:57 2025 -0500

    fix build script

commit d7f2d75f6a05183d116e4ccfebda1b5afde210bd
Author: Rafael Singer <rafaelsinger66@gmail.com>
Date:   Thu Apr 17 14:46:05 2025 -0500

    add new build script

+++ src/app/api/users/[id]/route.ts
@@ -109,9 +109,11 @@ export const PUT = async (request: NextRequest, context: { params: Promise<{ id:
         industryId: industryId,
         lookingForRoommate: userData.lookingForRoommate,
         visibilityOptions: userData.visibilityOptions,
+        classYear: userData.classYear,
       },
       include: {
         location: true,
+        industry: true,
       },
     });
 
@@ -144,6 +146,22 @@ export const DELETE = async (
       return new NextResponse('Unauthorized - Cannot delete another user account', { status: 401 });
     }
 
+    // First check if the user exists
+    const userExists = await prisma.user.findUnique({
+      where: {
+        id: id,
+      },
+    });
+
+    if (!userExists) {
+      return new NextResponse(JSON.stringify({ success: false, message: 'User not found' }), {
+        status: 404,
+        headers: {
+          'Content-Type': 'application/json',
+        },
+      });
+    }
+
     // Delete user account (Prisma will cascade delete related data thanks to our schema setup)
     await prisma.user.delete({
       where: {
@@ -160,8 +178,38 @@ export const DELETE = async (
         },
       }
     );
-  } catch (error) {
+  } catch (error: unknown) {
     console.error('Error deleting user account:', error);
-    return new NextResponse('Internal Server Error', { status: 500 });
+
+    // Check if it's a Prisma error (more detailed error handling)
+    if (typeof error === 'object' && error !== null && 'code' in error && error.code === 'P2025') {
+      return new NextResponse(
+        JSON.stringify({
+          success: false,
+          message: 'Account not found or already deleted',
+          error: error instanceof Error ? error.message : String(error),
+        }),
+        {
+          status: 404,
+          headers: {
+            'Content-Type': 'application/json',
+          },
+        }
+      );
+    }
+
+    return new NextResponse(
+      JSON.stringify({
+        success: false,
+        message: 'Error deleting account',
+        error: error instanceof Error ? error.message : String(error),
+      }),
+      {
+        status: 500,
+        headers: {
+          'Content-Type': 'application/json',
+        },
+      }
+    );
   }
 };
